chore(auth): split oversized policy test files; docs + version bump (BR-7085)
- Split condition.rs (1093) and evaluator.rs (843) test modules into nested
  file-based submodules; all .rs files now under 500 lines.
- Bump workspace version 0.1.5 -> 0.1.6.
- Add troubleshooting entry + upgrade-manual behavior-change note for the
  multivalued-condition-key fix.
- Refresh backlog.md (f64 numeric equality, non-2012-10-17 Version, policy-
  variable divergence) and regenerate todo-index.md.

Refs: BR-7085
