+++ crates/auth/src/policy/condition.rs
@@ -319,775 +319,4 @@ fn parse_epoch_millis(s: &str) -> Option<i128> {
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::policy::context::ConditionContext;
-    use std::collections::HashMap;
-
-    /// Simple test context that maps keys to values.
-    struct TestContext(HashMap<String, Vec<String>>);
-
-    impl TestContext {
-        fn new() -> Self {
-            Self(HashMap::new())
-        }
-
-        fn with(mut self, key: &str, values: Vec<&str>) -> Self {
-            self.0.insert(
-                key.to_owned(),
-                values.into_iter().map(ToOwned::to_owned).collect(),
-            );
-            self
-        }
-    }
-
-    impl ConditionContext for TestContext {
-        fn resolve_key(&self, key: &str) -> Option<Vec<&str>> {
-            self.0
-                .get(key)
-                .map(|v| v.iter().map(|s| s.as_str()).collect())
-        }
-
-        fn is_multivalued_key(&self, key: &str) -> bool {
-            // Mirror the real RequestContext classification so tests exercise the
-            // production arity rule.
-            matches!(key, "dynamodb:LeadingKeys" | "dynamodb:Attributes")
-        }
-    }
-
-    fn cond(op: ConditionOperator, key: &str, values: Vec<&str>) -> Condition {
-        Condition {
-            operator: op,
-            key: key.to_owned(),
-            values: values.into_iter().map(ToOwned::to_owned).collect(),
-        }
-    }
-
-    // --- StringEquals ---
-
-    #[test]
-    fn string_equals_match() {
-        let ctx = TestContext::new().with("k", vec!["hello"]);
-        assert!(evaluate_condition(
-            &cond(ConditionOperator::StringEquals, "k", vec!["hello"]),
-            &ctx
-        ));
-    }
-
-    #[test]
-    fn string_equals_no_match() {
-        let ctx = TestContext::new().with("k", vec!["hello"]);
-        assert!(!evaluate_condition(
-            &cond(ConditionOperator::StringEquals, "k", vec!["world"]),
-            &ctx
-        ));
-    }
-
-    #[test]
-    fn string_equals_absent_key() {
-        let ctx = TestContext::new();
-        assert!(!evaluate_condition(
-            &cond(ConditionOperator::StringEquals, "k", vec!["hello"]),
-            &ctx
-        ));
-    }
-
-    #[test]
-    fn string_equals_multiple_policy_values() {
-        let ctx = TestContext::new().with("k", vec!["b"]);
-        assert!(evaluate_condition(
-            &cond(ConditionOperator::StringEquals, "k", vec!["a", "b", "c"]),
-            &ctx
-        ));
-    }
-
-    // --- StringLike ---
-
-    #[test]
-    fn string_like_wildcard() {
-        let ctx = TestContext::new().with("k", vec!["hello-world"]);
-        assert!(evaluate_condition(
-            &cond(ConditionOperator::StringLike, "k", vec!["hello-*"]),
-            &ctx
-        ));
-    }
-
-    #[test]
-    fn string_not_like() {
-        let ctx = TestContext::new().with("k", vec!["hello"]);
-        assert!(evaluate_condition(
-            &cond(ConditionOperator::StringNotLike, "k", vec!["world*"]),
-            &ctx
-        ));
-    }
-
-    // --- StringEqualsIgnoreCase ---
-
-    #[test]
-    fn string_equals_ignore_case() {
-        let ctx = TestContext::new().with("k", vec!["Hello"]);
-        assert!(evaluate_condition(
-            &cond(
-                ConditionOperator::StringEqualsIgnoreCase,
-                "k",
-                vec!["hello"]
-            ),
-            &ctx
-        ));
-    }
-
-    // --- Numeric ---
-
-    #[test]
-    fn numeric_equals() {
-        let ctx = TestContext::new().with("k", vec!["42"]);
-        assert!(evaluate_condition(
-            &cond(ConditionOperator::NumericEquals, "k", vec!["42"]),
-            &ctx
-        ));
-    }
-
-    #[test]
-    fn numeric_less_than() {
-        let ctx = TestContext::new().with("k", vec!["5"]);
-        assert!(evaluate_condition(
-            &cond(ConditionOperator::NumericLessThan, "k", vec!["10"]),
-            &ctx
-        ));
-    }
-
-    #[test]
-    fn numeric_greater_than_equals() {
-        let ctx = TestContext::new().with("k", vec!["10"]);
-        assert!(evaluate_condition(
-            &cond(ConditionOperator::NumericGreaterThanEquals, "k", vec!["10"]),
-            &ctx
-        ));
-    }
-
-    #[test]
-    fn numeric_invalid_parse() {
-        let ctx = TestContext::new().with("k", vec!["abc"]);
-        assert!(!evaluate_condition(
-            &cond(ConditionOperator::NumericEquals, "k", vec!["42"]),
-            &ctx
-        ));
-    }
-
-    // --- Bool ---
-
-    #[test]
-    fn bool_true() {
-        let ctx = TestContext::new().with("k", vec!["true"]);
-        assert!(evaluate_condition(
-            &cond(ConditionOperator::Bool, "k", vec!["true"]),
-            &ctx
-        ));
-    }
-
-    #[test]
-    fn bool_case_insensitive() {
-        let ctx = TestContext::new().with("k", vec!["True"]);
-        assert!(evaluate_condition(
-            &cond(ConditionOperator::Bool, "k", vec!["true"]),
-            &ctx
-        ));
-    }
-
-    // --- Null ---
-
-    #[test]
-    fn null_true_absent_key() {
-        let ctx = TestContext::new();
-        assert!(evaluate_condition(
-            &cond(ConditionOperator::Null, "k", vec!["true"]),
-            &ctx
-        ));
-    }
-
-    #[test]
-    fn null_true_present_key() {
-        let ctx = TestContext::new().with("k", vec!["val"]);
-        assert!(!evaluate_condition(
-            &cond(ConditionOperator::Null, "k", vec!["true"]),
-            &ctx
-        ));
-    }
-
-    #[test]
-    fn null_false_present_key() {
-        let ctx = TestContext::new().with("k", vec!["val"]);
-        assert!(evaluate_condition(
-            &cond(ConditionOperator::Null, "k", vec!["false"]),
-            &ctx
-        ));
-    }
-
-    #[test]
-    fn null_false_absent_key() {
-        let ctx = TestContext::new();
-        assert!(!evaluate_condition(
-            &cond(ConditionOperator::Null, "k", vec!["false"]),
-            &ctx
-        ));
-    }
-
-    // --- ArnLike ---
-
-    #[test]
-    fn arn_like_match() {
-        let ctx = TestContext::new().with("k", vec!["arn:aws:dynamodb:us-east-1:123:table/Users"]);
-        assert!(evaluate_condition(
-            &cond(
-                ConditionOperator::ArnLike,
-                "k",
-                vec!["arn:aws:dynamodb:*:*:table/User*"]
-            ),
-            &ctx
-        ));
-    }
-
-    #[test]
-    fn arn_not_like() {
-        let ctx = TestContext::new().with("k", vec!["arn:aws:dynamodb:us-east-1:123:table/Orders"]);
-        assert!(evaluate_condition(
-            &cond(
-                ConditionOperator::ArnNotLike,
-                "k",
-                vec!["arn:aws:dynamodb:*:*:table/User*"]
-            ),
-            &ctx
-        ));
-    }
-
-    // --- IfExists ---
-
-    #[test]
-    fn if_exists_absent_key_passes() {
-        let ctx = TestContext::new();
-        assert!(evaluate_condition(
-            &cond(
-                ConditionOperator::IfExists(Box::new(ConditionOperator::StringEquals)),
-                "k",
-                vec!["val"]
-            ),
-            &ctx
-        ));
-    }
-
-    #[test]
-    fn if_exists_present_key_evaluates() {
-        let ctx = TestContext::new().with("k", vec!["val"]);
-        assert!(evaluate_condition(
-            &cond(
-                ConditionOperator::IfExists(Box::new(ConditionOperator::StringEquals)),
-                "k",
-                vec!["val"]
-            ),
-            &ctx
-        ));
-    }
-
-    #[test]
-    fn if_exists_present_key_fails() {
-        let ctx = TestContext::new().with("k", vec!["other"]);
-        assert!(!evaluate_condition(
-            &cond(
-                ConditionOperator::IfExists(Box::new(ConditionOperator::StringEquals)),
-                "k",
-                vec!["val"]
-            ),
-            &ctx
-        ));
-    }
-
-    // --- ForAllValues ---
-
-    #[test]
-    fn for_all_values_all_match() {
-        let ctx = TestContext::new().with("k", vec!["a", "b"]);
-        assert!(evaluate_condition(
-            &cond(
-                ConditionOperator::ForAllValues(Box::new(ConditionOperator::StringEquals)),
-                "k",
-                vec!["a", "b", "c"]
-            ),
-            &ctx
-        ));
-    }
-
-    #[test]
-    fn for_all_values_one_missing() {
-        let ctx = TestContext::new().with("k", vec!["a", "d"]);
-        assert!(!evaluate_condition(
-            &cond(
-                ConditionOperator::ForAllValues(Box::new(ConditionOperator::StringEquals)),
-                "k",
-                vec!["a", "b", "c"]
-            ),
-            &ctx
-        ));
-    }
-
-    #[test]
-    fn for_all_values_absent_key_vacuously_true() {
-        let ctx = TestContext::new();
-        assert!(evaluate_condition(
-            &cond(
-                ConditionOperator::ForAllValues(Box::new(ConditionOperator::StringEquals)),
-                "k",
-                vec!["a"]
-            ),
-            &ctx
-        ));
-    }
-
-    // --- ForAnyValue ---
-
-    #[test]
-    fn for_any_value_one_match() {
-        let ctx = TestContext::new().with("k", vec!["x", "a"]);
-        assert!(evaluate_condition(
-            &cond(
-                ConditionOperator::ForAnyValue(Box::new(ConditionOperator::StringEquals)),
-                "k",
-                vec!["a"]
-            ),
-            &ctx
-        ));
-    }
-
-    #[test]
-    fn for_any_value_no_match() {
-        let ctx = TestContext::new().with("k", vec!["x", "y"]);
-        assert!(!evaluate_condition(
-            &cond(
-                ConditionOperator::ForAnyValue(Box::new(ConditionOperator::StringEquals)),
-                "k",
-                vec!["a"]
-            ),
-            &ctx
-        ));
-    }
-
-    #[test]
-    fn for_any_value_absent_key_false() {
-        let ctx = TestContext::new();
-        assert!(!evaluate_condition(
-            &cond(
-                ConditionOperator::ForAnyValue(Box::new(ConditionOperator::StringEquals)),
-                "k",
-                vec!["a"]
-            ),
-            &ctx
-        ));
-    }
-
-    #[test]
-    fn for_any_value_if_exists_absent_key_true() {
-        let ctx = TestContext::new();
-        assert!(evaluate_condition(
-            &cond(
-                ConditionOperator::ForAnyValue(Box::new(ConditionOperator::IfExists(Box::new(
-                    ConditionOperator::StringEquals
-                )))),
-                "k",
-                vec!["a"]
-            ),
-            &ctx
-        ));
-    }
-
-    // --- Date operators ---
-
-    #[test]
-    fn date_equals() {
-        let ctx = TestContext::new().with("k", vec!["2026-01-01T00:00:00Z"]);
-        assert!(evaluate_condition(
-            &cond(
-                ConditionOperator::DateEquals,
-                "k",
-                vec!["2026-01-01T00:00:00Z"]
-            ),
-            &ctx
-        ));
-    }
-
-    #[test]
-    fn date_less_than() {
-        let ctx = TestContext::new().with("k", vec!["2025-01-01T00:00:00Z"]);
-        assert!(evaluate_condition(
-            &cond(
-                ConditionOperator::DateLessThan,
-                "k",
-                vec!["2026-01-01T00:00:00Z"]
-            ),
-            &ctx
-        ));
-    }
-
-    // --- BR-7085: bare operators on multivalued keys must never match ---
-    //
-    // Verified against real AWS IAM (Policy Simulator + end-to-end DynamoDB): a bare
-    // (non-ForAllValues/ForAnyValue) operator applied to a multivalued key
-    // (dynamodb:Attributes, dynamodb:LeadingKeys) never matches, regardless of the
-    // operator or how many values are present. Only the set qualifiers match.
-
-    #[test]
-    fn bare_op_multivalued_single_value_no_match() {
-        // Even a single requested attribute equal to the policy value does NOT match
-        // (real AWS returns "allowed"). This is the reporter's "control test", which
-        // AWS does not deny.
-        let ctx = TestContext::new().with("dynamodb:Attributes", vec!["ssn"]);
-        assert!(!evaluate_condition(
-            &cond(
-                ConditionOperator::StringEquals,
-                "dynamodb:Attributes",
-                vec!["ssn"]
-            ),
-            &ctx
-        ));
-    }
-
-    #[test]
-    fn bare_op_multivalued_all_match_no_match() {
-        let ctx = TestContext::new().with("dynamodb:Attributes", vec!["ssn", "salary"]);
-        assert!(!evaluate_condition(
-            &cond(
-                ConditionOperator::StringEquals,
-                "dynamodb:Attributes",
-                vec!["ssn", "salary"]
-            ),
-            &ctx
-        ));
-    }
-
-    #[test]
-    fn bare_op_multivalued_extra_value_no_match() {
-        // The reported "bypass" shape: request ssn + fullname, policy denies [ssn].
-        let ctx = TestContext::new().with("dynamodb:Attributes", vec!["ssn", "fullname"]);
-        assert!(!evaluate_condition(
-            &cond(
-                ConditionOperator::StringEquals,
-                "dynamodb:Attributes",
-                vec!["ssn"]
-            ),
-            &ctx
-        ));
-    }
-
-    #[test]
-    fn bare_negative_op_multivalued_no_match() {
-        let ctx = TestContext::new().with("dynamodb:Attributes", vec!["fullname"]);
-        assert!(!evaluate_condition(
-            &cond(
-                ConditionOperator::StringNotEquals,
-                "dynamodb:Attributes",
-                vec!["ssn"]
-            ),
-            &ctx
-        ));
-    }
-
-    #[test]
-    fn bare_string_like_multivalued_no_match() {
-        let ctx = TestContext::new().with("dynamodb:Attributes", vec!["ssn"]);
-        assert!(!evaluate_condition(
-            &cond(ConditionOperator::StringLike, "dynamodb:Attributes", vec!["ss*"]),
-            &ctx
-        ));
-    }
-
-    #[test]
-    fn bare_op_multivalued_leading_keys_no_match() {
-        let ctx = TestContext::new().with("dynamodb:LeadingKeys", vec!["user1"]);
-        assert!(!evaluate_condition(
-            &cond(
-                ConditionOperator::StringEquals,
-                "dynamodb:LeadingKeys",
-                vec!["user1"]
-            ),
-            &ctx
-        ));
-    }
-
-    #[test]
-    fn if_exists_bare_op_multivalued_present_no_match() {
-        let ctx = TestContext::new().with("dynamodb:Attributes", vec!["ssn"]);
-        assert!(!evaluate_condition(
-            &cond(
-                ConditionOperator::IfExists(Box::new(ConditionOperator::StringEquals)),
-                "dynamodb:Attributes",
-                vec!["ssn"]
-            ),
-            &ctx
-        ));
-    }
-
-    #[test]
-    fn if_exists_multivalued_absent_passes() {
-        // IfExists still passes when the key is absent.
-        let ctx = TestContext::new();
-        assert!(evaluate_condition(
-            &cond(
-                ConditionOperator::IfExists(Box::new(ConditionOperator::StringEquals)),
-                "dynamodb:Attributes",
-                vec!["ssn"]
-            ),
-            &ctx
-        ));
-    }
-
-    #[test]
-    fn for_any_value_multivalued_deny_matches() {
-        // The CORRECT deny pattern: ForAnyValue:StringEquals fires when ANY requested
-        // attribute is in the denied set.
-        let ctx = TestContext::new().with("dynamodb:Attributes", vec!["ssn", "fullname"]);
-        assert!(evaluate_condition(
-            &cond(
-                ConditionOperator::ForAnyValue(Box::new(ConditionOperator::StringEquals)),
-                "dynamodb:Attributes",
-                vec!["ssn"]
-            ),
-            &ctx
-        ));
-    }
-
-    #[test]
-    fn for_all_values_multivalued_allowlist() {
-        // Allowlist pattern: ForAllValues:StringEquals allows only when EVERY requested
-        // attribute is in the allowed set.
-        let ctx_ok = TestContext::new().with("dynamodb:Attributes", vec!["pk", "fullname"]);
-        assert!(evaluate_condition(
-            &cond(
-                ConditionOperator::ForAllValues(Box::new(ConditionOperator::StringEquals)),
-                "dynamodb:Attributes",
-                vec!["pk", "fullname"]
-            ),
-            &ctx_ok
-        ));
-        let ctx_bad = TestContext::new().with("dynamodb:Attributes", vec!["pk", "ssn"]);
-        assert!(!evaluate_condition(
-            &cond(
-                ConditionOperator::ForAllValues(Box::new(ConditionOperator::StringEquals)),
-                "dynamodb:Attributes",
-                vec!["pk", "fullname"]
-            ),
-            &ctx_bad
-        ));
-    }
-
-    #[test]
-    fn bare_op_single_valued_key_still_matches() {
-        // Regression: bare operators on single-valued keys are unaffected.
-        let ctx = TestContext::new().with("dynamodb:Select", vec!["ALL_ATTRIBUTES"]);
-        assert!(evaluate_condition(
-            &cond(
-                ConditionOperator::StringEquals,
-                "dynamodb:Select",
-                vec!["ALL_ATTRIBUTES"]
-            ),
-            &ctx
-        ));
-    }
-
-    // --- Negative operators with multiple policy values ---
-
-    #[test]
-    fn string_not_equals_single_policy_value() {
-        let ctx = TestContext::new().with("k", vec!["admin"]);
-        // "admin" == "admin", so StringNotEquals should be false
-        assert!(!evaluate_condition(
-            &cond(ConditionOperator::StringNotEquals, "k", vec!["admin"]),
-            &ctx
-        ));
-    }
-
-    #[test]
-    fn string_not_equals_multiple_policy_values_match() {
-        let ctx = TestContext::new().with("k", vec!["admin"]);
-        // "admin" is in {"admin", "root"}, so StringNotEquals should be false
-        assert!(!evaluate_condition(
-            &cond(
-                ConditionOperator::StringNotEquals,
-                "k",
-                vec!["admin", "root"]
-            ),
-            &ctx
-        ));
-    }
-
-    #[test]
-    fn string_not_equals_multiple_policy_values_no_match() {
-        let ctx = TestContext::new().with("k", vec!["user"]);
-        // "user" is NOT in {"admin", "root"}, so StringNotEquals should be true
-        assert!(evaluate_condition(
-            &cond(
-                ConditionOperator::StringNotEquals,
-                "k",
-                vec!["admin", "root"]
-            ),
-            &ctx
-        ));
-    }
-
-    #[test]
-    fn string_not_like_multiple_policy_values() {
-        let ctx = TestContext::new().with("k", vec!["hello-world"]);
-        // "hello-world" matches "hello-*", so StringNotLike should be false
-        assert!(!evaluate_condition(
-            &cond(
-                ConditionOperator::StringNotLike,
-                "k",
-                vec!["hello-*", "foo-*"]
-            ),
-            &ctx
-        ));
-    }
-
-    #[test]
-    fn string_not_like_multiple_policy_values_no_match() {
-        let ctx = TestContext::new().with("k", vec!["bar-baz"]);
-        // "bar-baz" matches neither "hello-*" nor "foo-*"
-        assert!(evaluate_condition(
-            &cond(
-                ConditionOperator::StringNotLike,
-                "k",
-                vec!["hello-*", "foo-*"]
-            ),
-            &ctx
-        ));
-    }
-
-    // --- Policy variable expansion ---
-
-    #[test]
-    fn policy_variable_expansion_principal_tag() {
-        let ctx = TestContext::new()
-            .with("dynamodb:ResourceTag/Team", vec!["Alpha"])
-            .with("aws:PrincipalTag/Team", vec!["Alpha"]);
-        assert!(evaluate_condition(
-            &cond(
-                ConditionOperator::StringEquals,
-                "dynamodb:ResourceTag/Team",
-                vec!["${aws:PrincipalTag/Team}"]
-            ),
-            &ctx
-        ));
-    }
-
-    #[test]
-    fn policy_variable_expansion_mismatch() {
-        let ctx = TestContext::new()
-            .with("dynamodb:ResourceTag/Team", vec!["Beta"])
-            .with("aws:PrincipalTag/Team", vec!["Alpha"]);
-        assert!(!evaluate_condition(
-            &cond(
-                ConditionOperator::StringEquals,
-                "dynamodb:ResourceTag/Team",
-                vec!["${aws:PrincipalTag/Team}"]
-            ),
-            &ctx
-        ));
-    }
-
-    #[test]
-    fn policy_variable_no_expansion_needed() {
-        let ctx = TestContext::new().with("k", vec!["hello"]);
-        assert!(evaluate_condition(
-            &cond(ConditionOperator::StringEquals, "k", vec!["hello"]),
-            &ctx
-        ));
-    }
-
-    #[test]
-    fn policy_variable_unresolvable_left_literal() {
-        let ctx = TestContext::new().with("k", vec!["${aws:PrincipalTag/Missing}"]);
-        // Unresolvable variable stays literal — matches if context has the literal
-        assert!(evaluate_condition(
-            &cond(
-                ConditionOperator::StringEquals,
-                "k",
-                vec!["${aws:PrincipalTag/Missing}"]
-            ),
-            &ctx
-        ));
-    }
-
-    // --- ForAllValues with negative operators ---
-
-    #[test]
-    fn for_all_values_string_not_equals_context_in_policy_set() {
-        // ForAllValues:StringNotEquals with context ["admin"] and policy ["admin", "root"]
-        // "admin" IS in the set, so the condition should be FALSE (deny applies).
-        let ctx = TestContext::new().with("k", vec!["admin"]);
-        assert!(!evaluate_condition(
-            &cond(
-                ConditionOperator::ForAllValues(Box::new(ConditionOperator::StringNotEquals)),
-                "k",
-                vec!["admin", "root"]
-            ),
-            &ctx
-        ));
-    }
-
-    #[test]
-    fn for_all_values_string_not_equals_context_not_in_policy_set() {
-        // ForAllValues:StringNotEquals with context ["user"] and policy ["admin", "root"]
-        // "user" is NOT in the set, so the condition should be TRUE.
-        let ctx = TestContext::new().with("k", vec!["user"]);
-        assert!(evaluate_condition(
-            &cond(
-                ConditionOperator::ForAllValues(Box::new(ConditionOperator::StringNotEquals)),
-                "k",
-                vec!["admin", "root"]
-            ),
-            &ctx
-        ));
-    }
-
-    #[test]
-    fn for_all_values_string_not_equals_mixed_context() {
-        // ForAllValues:StringNotEquals with context ["user", "admin"] and policy ["admin", "root"]
-        // "admin" IS in the set, so the condition should be FALSE.
-        let ctx = TestContext::new().with("k", vec!["user", "admin"]);
-        assert!(!evaluate_condition(
-            &cond(
-                ConditionOperator::ForAllValues(Box::new(ConditionOperator::StringNotEquals)),
-                "k",
-                vec!["admin", "root"]
-            ),
-            &ctx
-        ));
-    }
-
-    #[test]
-    fn for_any_value_string_not_equals_all_in_set() {
-        // ForAnyValue:StringNotEquals with context ["admin", "root"] and policy ["admin", "root"]
-        // Both context values are in the policy set, so none satisfy "not in set" → FALSE.
-        let ctx = TestContext::new().with("k", vec!["admin", "root"]);
-        assert!(!evaluate_condition(
-            &cond(
-                ConditionOperator::ForAnyValue(Box::new(ConditionOperator::StringNotEquals)),
-                "k",
-                vec!["admin", "root"]
-            ),
-            &ctx
-        ));
-    }
-
-    #[test]
-    fn for_any_value_string_not_equals_one_outside_set() {
-        // ForAnyValue:StringNotEquals with context ["admin", "user"] and policy ["admin", "root"]
-        // "user" is NOT in the set, so at least one satisfies "not in set" → TRUE.
-        let ctx = TestContext::new().with("k", vec!["admin", "user"]);
-        assert!(evaluate_condition(
-            &cond(
-                ConditionOperator::ForAnyValue(Box::new(ConditionOperator::StringNotEquals)),
-                "k",
-                vec!["admin", "root"]
-            ),
-            &ctx
-        ));
-    }
-}
+mod tests;
