+++ crates/auth/src/policy/evaluator.rs
@@ -185,659 +185,4 @@ fn conditions_match(statement: &Statement, context: &impl ConditionContext) -> b
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::policy::context::ConditionContext;
-    use crate::policy::document::PolicyDocument;
-    use std::collections::HashMap;
-
-    /// Simple test context.
-    struct Ctx(HashMap<String, Vec<String>>);
-
-    impl Ctx {
-        fn empty() -> Self {
-            Self(HashMap::new())
-        }
-
-        fn with(mut self, key: &str, values: Vec<&str>) -> Self {
-            self.0.insert(
-                key.to_owned(),
-                values.into_iter().map(ToOwned::to_owned).collect(),
-            );
-            self
-        }
-    }
-
-    impl ConditionContext for Ctx {
-        fn resolve_key(&self, key: &str) -> Option<Vec<&str>> {
-            self.0
-                .get(key)
-                .map(|v| v.iter().map(|s| s.as_str()).collect())
-        }
-
-        fn is_multivalued_key(&self, key: &str) -> bool {
-            matches!(key, "dynamodb:LeadingKeys" | "dynamodb:Attributes")
-        }
-    }
-
-    fn parse(json: &str) -> PolicyDocument {
-        PolicyDocument::from_json(json).unwrap()
-    }
-
-    // --- Basic Allow/Deny ---
-
-    #[test]
-    fn simple_allow() {
-        let policy = parse(
-            r#"{"Version":"2012-10-17","Statement":[{
-                "Effect":"Allow","Action":"dynamodb:PutItem","Resource":"*"
-            }]}"#,
-        );
-        assert_eq!(
-            evaluate_policies(
-                &[policy],
-                None,
-                None,
-                "dynamodb:PutItem",
-                "arn:aws:dynamodb:us-east-1:123:table/T",
-                &Ctx::empty()
-            ),
-            AuthzDecision::Allow
-        );
-    }
-
-    #[test]
-    fn implicit_deny_no_matching_allow() {
-        let policy = parse(
-            r#"{"Version":"2012-10-17","Statement":[{
-                "Effect":"Allow","Action":"dynamodb:GetItem","Resource":"*"
-            }]}"#,
-        );
-        assert_eq!(
-            evaluate_policies(
-                &[policy],
-                None,
-                None,
-                "dynamodb:PutItem",
-                "arn:aws:dynamodb:us-east-1:123:table/T",
-                &Ctx::empty()
-            ),
-            AuthzDecision::Deny
-        );
-    }
-
-    #[test]
-    fn explicit_deny_overrides_allow() {
-        let allow = parse(
-            r#"{"Version":"2012-10-17","Statement":[{
-                "Effect":"Allow","Action":"dynamodb:*","Resource":"*"
-            }]}"#,
-        );
-        let deny = parse(
-            r#"{"Version":"2012-10-17","Statement":[{
-                "Effect":"Deny","Action":"dynamodb:DeleteTable","Resource":"*"
-            }]}"#,
-        );
-        assert_eq!(
-            evaluate_policies(
-                &[allow, deny],
-                None,
-                None,
-                "dynamodb:DeleteTable",
-                "arn:aws:dynamodb:us-east-1:123:table/T",
-                &Ctx::empty()
-            ),
-            AuthzDecision::Deny
-        );
-    }
-
-    // --- Case-insensitive action matching ---
-
-    #[test]
-    fn action_matching_is_case_insensitive() {
-        let policy = parse(
-            r#"{"Version":"2012-10-17","Statement":[{
-                "Effect":"Allow","Action":"dynamodb:putitem","Resource":"*"
-            }]}"#,
-        );
-        assert_eq!(
-            evaluate_policies(
-                &[policy],
-                None,
-                None,
-                "dynamodb:PutItem",
-                "arn:aws:dynamodb:us-east-1:123:table/T",
-                &Ctx::empty()
-            ),
-            AuthzDecision::Allow
-        );
-    }
-
-    #[test]
-    fn deny_case_insensitive() {
-        let allow = parse(
-            r#"{"Version":"2012-10-17","Statement":[{
-                "Effect":"Allow","Action":"dynamodb:*","Resource":"*"
-            }]}"#,
-        );
-        let deny = parse(
-            r#"{"Version":"2012-10-17","Statement":[{
-                "Effect":"Deny","Action":"dynamodb:deletetable","Resource":"*"
-            }]}"#,
-        );
-        assert_eq!(
-            evaluate_policies(
-                &[allow, deny],
-                None,
-                None,
-                "dynamodb:DeleteTable",
-                "arn:aws:dynamodb:us-east-1:123:table/T",
-                &Ctx::empty()
-            ),
-            AuthzDecision::Deny
-        );
-    }
-
-    // --- NotAction ---
-
-    #[test]
-    fn not_action_deny() {
-        let policy = parse(
-            r#"{"Version":"2012-10-17","Statement":[{
-                "Effect":"Deny","NotAction":["dynamodb:GetItem"],"Resource":"*"
-            }]}"#,
-        );
-        // PutItem is not in the NotAction list, so the Deny applies
-        assert_eq!(
-            evaluate_policies(
-                &[policy],
-                None,
-                None,
-                "dynamodb:PutItem",
-                "arn:aws:dynamodb:us-east-1:123:table/T",
-                &Ctx::empty()
-            ),
-            AuthzDecision::Deny
-        );
-    }
-
-    #[test]
-    fn not_action_deny_excluded_action_not_denied() {
-        let deny = parse(
-            r#"{"Version":"2012-10-17","Statement":[{
-                "Effect":"Deny","NotAction":["dynamodb:GetItem"],"Resource":"*"
-            }]}"#,
-        );
-        let allow = parse(
-            r#"{"Version":"2012-10-17","Statement":[{
-                "Effect":"Allow","Action":"dynamodb:GetItem","Resource":"*"
-            }]}"#,
-        );
-        // GetItem is excluded from the Deny, and explicitly allowed
-        assert_eq!(
-            evaluate_policies(
-                &[deny, allow],
-                None,
-                None,
-                "dynamodb:GetItem",
-                "arn:aws:dynamodb:us-east-1:123:table/T",
-                &Ctx::empty()
-            ),
-            AuthzDecision::Allow
-        );
-    }
-
-    // --- NotResource ---
-
-    #[test]
-    fn not_resource_deny() {
-        let deny = parse(
-            r#"{"Version":"2012-10-17","Statement":[{
-                "Effect":"Deny","Action":"dynamodb:*",
-                "NotResource":["arn:aws:dynamodb:*:*:table/AllowedTable"]
-            }]}"#,
-        );
-        let allow = parse(
-            r#"{"Version":"2012-10-17","Statement":[{
-                "Effect":"Allow","Action":"dynamodb:*","Resource":"*"
-            }]}"#,
-        );
-        // Access to AllowedTable is not denied (excluded from NotResource)
-        assert_eq!(
-            evaluate_policies(
-                &[deny.clone(), allow.clone()],
-                None,
-                None,
-                "dynamodb:PutItem",
-                "arn:aws:dynamodb:us-east-1:123:table/AllowedTable",
-                &Ctx::empty()
-            ),
-            AuthzDecision::Allow
-        );
-        // Access to OtherTable is denied
-        assert_eq!(
-            evaluate_policies(
-                &[deny, allow],
-                None,
-                None,
-                "dynamodb:PutItem",
-                "arn:aws:dynamodb:us-east-1:123:table/OtherTable",
-                &Ctx::empty()
-            ),
-            AuthzDecision::Deny
-        );
-    }
-
-    // --- Permissions Boundary ---
-
-    #[test]
-    fn boundary_restricts_allow() {
-        let identity = parse(
-            r#"{"Version":"2012-10-17","Statement":[{
-                "Effect":"Allow","Action":"dynamodb:*","Resource":"*"
-            }]}"#,
-        );
-        let boundary = parse(
-            r#"{"Version":"2012-10-17","Statement":[{
-                "Effect":"Allow","Action":"dynamodb:GetItem","Resource":"*"
-            }]}"#,
-        );
-        // Identity allows all, but boundary only allows GetItem
-        assert_eq!(
-            evaluate_policies(
-                std::slice::from_ref(&identity),
-                Some(&boundary),
-                None,
-                "dynamodb:GetItem",
-                "arn:aws:dynamodb:us-east-1:123:table/T",
-                &Ctx::empty()
-            ),
-            AuthzDecision::Allow
-        );
-        assert_eq!(
-            evaluate_policies(
-                &[identity],
-                Some(&boundary),
-                None,
-                "dynamodb:PutItem",
-                "arn:aws:dynamodb:us-east-1:123:table/T",
-                &Ctx::empty()
-            ),
-            AuthzDecision::Deny
-        );
-    }
-
-    #[test]
-    fn boundary_deny_overrides() {
-        let identity = parse(
-            r#"{"Version":"2012-10-17","Statement":[{
-                "Effect":"Allow","Action":"dynamodb:*","Resource":"*"
-            }]}"#,
-        );
-        let boundary = parse(
-            r#"{"Version":"2012-10-17","Statement":[
-                {"Effect":"Allow","Action":"dynamodb:*","Resource":"*"},
-                {"Effect":"Deny","Action":"dynamodb:DeleteTable","Resource":"*"}
-            ]}"#,
-        );
-        assert_eq!(
-            evaluate_policies(
-                &[identity],
-                Some(&boundary),
-                None,
-                "dynamodb:DeleteTable",
-                "arn:aws:dynamodb:us-east-1:123:table/T",
-                &Ctx::empty()
-            ),
-            AuthzDecision::Deny
-        );
-    }
-
-    // --- Session Policy ---
-
-    #[test]
-    fn session_policy_restricts() {
-        let role = parse(
-            r#"{"Version":"2012-10-17","Statement":[{
-                "Effect":"Allow","Action":"dynamodb:*","Resource":"*"
-            }]}"#,
-        );
-        let session = parse(
-            r#"{"Version":"2012-10-17","Statement":[{
-                "Effect":"Allow","Action":"dynamodb:GetItem","Resource":"*"
-            }]}"#,
-        );
-        assert_eq!(
-            evaluate_policies(
-                std::slice::from_ref(&role),
-                None,
-                Some(&session),
-                "dynamodb:GetItem",
-                "arn:aws:dynamodb:us-east-1:123:table/T",
-                &Ctx::empty()
-            ),
-            AuthzDecision::Allow
-        );
-        assert_eq!(
-            evaluate_policies(
-                &[role],
-                None,
-                Some(&session),
-                "dynamodb:PutItem",
-                "arn:aws:dynamodb:us-east-1:123:table/T",
-                &Ctx::empty()
-            ),
-            AuthzDecision::Deny
-        );
-    }
-
-    // --- Conditions (ABAC) ---
-
-    #[test]
-    fn condition_tag_match_allows() {
-        let policy = parse(
-            r#"{"Version":"2012-10-17","Statement":[{
-                "Effect":"Allow","Action":"dynamodb:*","Resource":"*",
-                "Condition":{"StringEquals":{"aws:PrincipalTag/Department":"Eng"}}
-            }]}"#,
-        );
-        let ctx = Ctx::empty().with("aws:PrincipalTag/Department", vec!["Eng"]);
-        assert_eq!(
-            evaluate_policies(
-                &[policy],
-                None,
-                None,
-                "dynamodb:PutItem",
-                "arn:aws:dynamodb:us-east-1:123:table/T",
-                &ctx
-            ),
-            AuthzDecision::Allow
-        );
-    }
-
-    #[test]
-    fn condition_tag_mismatch_denies() {
-        let policy = parse(
-            r#"{"Version":"2012-10-17","Statement":[{
-                "Effect":"Allow","Action":"dynamodb:*","Resource":"*",
-                "Condition":{"StringEquals":{"aws:PrincipalTag/Department":"Eng"}}
-            }]}"#,
-        );
-        let ctx = Ctx::empty().with("aws:PrincipalTag/Department", vec!["Sales"]);
-        assert_eq!(
-            evaluate_policies(
-                &[policy],
-                None,
-                None,
-                "dynamodb:PutItem",
-                "arn:aws:dynamodb:us-east-1:123:table/T",
-                &ctx
-            ),
-            AuthzDecision::Deny
-        );
-    }
-
-    // --- Combined: boundary + session + conditions ---
-
-    #[test]
-    fn full_evaluation_all_phases() {
-        let identity = parse(
-            r#"{"Version":"2012-10-17","Statement":[{
-                "Effect":"Allow","Action":"dynamodb:GetItem","Resource":"*",
-                "Condition":{"StringEquals":{"aws:PrincipalTag/Team":"Alpha"}}
-            }]}"#,
-        );
-        let boundary = parse(
-            r#"{"Version":"2012-10-17","Statement":[{
-                "Effect":"Allow","Action":["dynamodb:GetItem","dynamodb:Query"],"Resource":"*"
-            }]}"#,
-        );
-        let session = parse(
-            r#"{"Version":"2012-10-17","Statement":[{
-                "Effect":"Allow","Action":"dynamodb:GetItem","Resource":"*"
-            }]}"#,
-        );
-        let ctx = Ctx::empty().with("aws:PrincipalTag/Team", vec!["Alpha"]);
-
-        // All phases pass
-        assert_eq!(
-            evaluate_policies(
-                std::slice::from_ref(&identity),
-                Some(&boundary),
-                Some(&session),
-                "dynamodb:GetItem",
-                "arn:aws:dynamodb:us-east-1:123:table/T",
-                &ctx
-            ),
-            AuthzDecision::Allow
-        );
-
-        // Wrong tag → identity policy condition fails
-        let ctx_wrong = Ctx::empty().with("aws:PrincipalTag/Team", vec!["Beta"]);
-        assert_eq!(
-            evaluate_policies(
-                &[identity],
-                Some(&boundary),
-                Some(&session),
-                "dynamodb:GetItem",
-                "arn:aws:dynamodb:us-east-1:123:table/T",
-                &ctx_wrong
-            ),
-            AuthzDecision::Deny
-        );
-    }
-
-    // --- Wildcard action matching ---
-
-    #[test]
-    fn wildcard_action() {
-        let policy = parse(
-            r#"{"Version":"2012-10-17","Statement":[{
-                "Effect":"Allow","Action":"dynamodb:*","Resource":"*"
-            }]}"#,
-        );
-        assert_eq!(
-            evaluate_policies(
-                &[policy],
-                None,
-                None,
-                "dynamodb:PutItem",
-                "arn:aws:dynamodb:us-east-1:123:table/T",
-                &Ctx::empty()
-            ),
-            AuthzDecision::Allow
-        );
-    }
-
-    // --- Resource ARN matching ---
-
-    #[test]
-    fn resource_arn_restricts() {
-        let policy = parse(
-            r#"{"Version":"2012-10-17","Statement":[{
-                "Effect":"Allow","Action":"dynamodb:*",
-                "Resource":"arn:aws:dynamodb:*:*:table/Users"
-            }]}"#,
-        );
-        assert_eq!(
-            evaluate_policies(
-                std::slice::from_ref(&policy),
-                None,
-                None,
-                "dynamodb:PutItem",
-                "arn:aws:dynamodb:us-east-1:123:table/Users",
-                &Ctx::empty()
-            ),
-            AuthzDecision::Allow
-        );
-        assert_eq!(
-            evaluate_policies(
-                &[policy],
-                None,
-                None,
-                "dynamodb:PutItem",
-                "arn:aws:dynamodb:us-east-1:123:table/Orders",
-                &Ctx::empty()
-            ),
-            AuthzDecision::Deny
-        );
-    }
-
-    // --- Multiple identity policies ---
-
-    #[test]
-    fn multiple_policies_any_allow() {
-        let read_only = parse(
-            r#"{"Version":"2012-10-17","Statement":[{
-                "Effect":"Allow","Action":"dynamodb:GetItem","Resource":"*"
-            }]}"#,
-        );
-        let write_only = parse(
-            r#"{"Version":"2012-10-17","Statement":[{
-                "Effect":"Allow","Action":"dynamodb:PutItem","Resource":"*"
-            }]}"#,
-        );
-        assert_eq!(
-            evaluate_policies(
-                &[read_only, write_only],
-                None,
-                None,
-                "dynamodb:PutItem",
-                "arn:aws:dynamodb:us-east-1:123:table/T",
-                &Ctx::empty()
-            ),
-            AuthzDecision::Allow
-        );
-    }
-
-    // --- Empty policies ---
-
-    #[test]
-    fn no_policies_implicit_deny() {
-        assert_eq!(
-            evaluate_policies(
-                &[],
-                None,
-                None,
-                "dynamodb:PutItem",
-                "arn:aws:dynamodb:us-east-1:123:table/T",
-                &Ctx::empty()
-            ),
-            AuthzDecision::Deny
-        );
-    }
-
-    // --- ForAllValues with leading keys (FGAC pattern) ---
-
-    #[test]
-    fn fgac_leading_keys() {
-        let policy = parse(
-            r#"{"Version":"2012-10-17","Statement":[{
-                "Effect":"Allow","Action":"dynamodb:*","Resource":"*",
-                "Condition":{
-                    "ForAllValues:StringEquals":{
-                        "dynamodb:LeadingKeys":["user-123"]
-                    }
-                }
-            }]}"#,
-        );
-        let ctx = Ctx::empty().with("dynamodb:LeadingKeys", vec!["user-123"]);
-        assert_eq!(
-            evaluate_policies(
-                std::slice::from_ref(&policy),
-                None,
-                None,
-                "dynamodb:GetItem",
-                "arn:aws:dynamodb:us-east-1:123:table/T",
-                &ctx
-            ),
-            AuthzDecision::Allow
-        );
-
-        let ctx_wrong = Ctx::empty().with("dynamodb:LeadingKeys", vec!["user-456"]);
-        assert_eq!(
-            evaluate_policies(
-                &[policy],
-                None,
-                None,
-                "dynamodb:GetItem",
-                "arn:aws:dynamodb:us-east-1:123:table/T",
-                &ctx_wrong
-            ),
-            AuthzDecision::Deny
-        );
-    }
-
-    // --- BR-7085: explicit-Deny handling for the dynamodb:Attributes FGAC pattern ---
-
-    fn fgac_policies() -> [PolicyDocument; 2] {
-        let allow = parse(
-            r#"{"Version":"2012-10-17","Statement":[{
-                "Effect":"Allow","Action":"dynamodb:*","Resource":"*"
-            }]}"#,
-        );
-        // Misconfigured denylist using a BARE StringEquals on the multivalued key.
-        let deny_bare = parse(
-            r#"{"Version":"2012-10-17","Statement":[{
-                "Effect":"Deny","Action":"dynamodb:*",
-                "Resource":"arn:aws:dynamodb:us-east-1:123:table/Employees",
-                "Condition":{"StringEquals":{"dynamodb:Attributes":["ssn"]}}
-            }]}"#,
-        );
-        [allow, deny_bare]
-    }
-
-    fn eval_attrs(policies: &[PolicyDocument], attrs: Vec<&str>) -> AuthzDecision {
-        let ctx = Ctx::empty().with("dynamodb:Attributes", attrs);
-        evaluate_policies(
-            policies,
-            None,
-            None,
-            "dynamodb:GetItem",
-            "arn:aws:dynamodb:us-east-1:123:table/Employees",
-            &ctx,
-        )
-    }
-
-    #[test]
-    fn bare_deny_on_attributes_is_a_noop_ssn_alone_allowed() {
-        // Matches real AWS: requesting ssn alone is ALLOWED (bare Deny never fires).
-        // ExtendDB previously DENIED this, creating a false sense of security.
-        let p = fgac_policies();
-        assert_eq!(eval_attrs(&p, vec!["ssn"]), AuthzDecision::Allow);
-    }
-
-    #[test]
-    fn bare_deny_on_attributes_is_a_noop_ssn_plus_fullname_allowed() {
-        let p = fgac_policies();
-        assert_eq!(eval_attrs(&p, vec!["ssn", "fullname"]), AuthzDecision::Allow);
-    }
-
-    #[test]
-    fn bare_deny_on_attributes_fullname_allowed() {
-        let p = fgac_policies();
-        assert_eq!(eval_attrs(&p, vec!["fullname"]), AuthzDecision::Allow);
-    }
-
-    #[test]
-    fn for_any_value_deny_on_attributes_is_effective() {
-        // The CORRECT pattern: ForAnyValue:StringEquals denies whenever ssn is requested,
-        // whether alone or alongside another attribute.
-        let allow = parse(
-            r#"{"Version":"2012-10-17","Statement":[{
-                "Effect":"Allow","Action":"dynamodb:*","Resource":"*"
-            }]}"#,
-        );
-        let deny_any = parse(
-            r#"{"Version":"2012-10-17","Statement":[{
-                "Effect":"Deny","Action":"dynamodb:*",
-                "Resource":"arn:aws:dynamodb:us-east-1:123:table/Employees",
-                "Condition":{"ForAnyValue:StringEquals":{"dynamodb:Attributes":["ssn"]}}
-            }]}"#,
-        );
-        let p = [allow, deny_any];
-        assert_eq!(eval_attrs(&p, vec!["ssn"]), AuthzDecision::Deny);
-        assert_eq!(eval_attrs(&p, vec!["ssn", "fullname"]), AuthzDecision::Deny);
-        assert_eq!(eval_attrs(&p, vec!["fullname"]), AuthzDecision::Allow);
-    }
-}
+mod tests;
